chore(login): refactor the user api folder
