+++ frontend/src/AppRoutes/Private.tsx
@@ -1,6 +1,6 @@
 import getLocalStorageApi from 'api/browser/localstorage/get';
 import setLocalStorageApi from 'api/browser/localstorage/set';
-import listUsers from 'api/user/listUsers';
+import listUsers from 'api/user/crud/getAll';
 import { FeatureKeys } from 'constants/features';
 import { LOCALSTORAGE } from 'constants/localStorage';
 import ROUTES from 'constants/routes';
