+++ frontend/src/api/user/crud/get.ts
@@ -2,13 +2,11 @@ import axios from 'api';
 import { ErrorResponseHandlerV2 } from 'api/ErrorResponseHandlerV2';
 import { AxiosError } from 'axios';
 import { ErrorV2Resp, SuccessResponseV2 } from 'types/api';
-import { PayloadProps, Props, UserResponse } from 'types/api/user/getUser';
+import { PayloadProps, UserResponse } from 'types/api/user/getUser';
 
-const getUser = async (
-	props: Props,
-): Promise<SuccessResponseV2<UserResponse>> => {
+const getUser = async (): Promise<SuccessResponseV2<UserResponse>> => {
 	try {
-		const response = await axios.get<PayloadProps>(`/user/${props.userId}`);
+		const response = await axios.get<PayloadProps>(`/user/me`);
 
 		return {
 			httpStatusCode: response.status,
