+++ frontend/src/api/user/factor_password/update.ts
