+++ frontend/src/api/user/invites/accept.ts
