+++ frontend/src/api/user/invites/createMany.ts
