+++ frontend/src/api/user/invites/delete.ts
