+++ frontend/src/container/AppLayout/index.tsx
@@ -9,8 +9,8 @@ import manageCreditCardApi from 'api/billing/manage';
 import getLocalStorageApi from 'api/browser/localstorage/get';
 import setLocalStorageApi from 'api/browser/localstorage/set';
 import logEvent from 'api/common/logEvent';
-import getUserLatestVersion from 'api/user/getLatestVersion';
-import getUserVersion from 'api/user/getVersion';
+import getUserLatestVersion from 'api/user/version/getLatestVersion';
+import getUserVersion from 'api/user/version/getVersion';
 import cx from 'classnames';
 import ChatSupportGateway from 'components/ChatSupportGateway/ChatSupportGateway';
 import OverlayScrollbar from 'components/OverlayScrollbar/OverlayScrollbar';
