+++ frontend/src/container/OrganizationSettings/Members/index.tsx
@@ -1,9 +1,8 @@
 import { Button, Modal, Space, Typography } from 'antd';
 import { ColumnsType } from 'antd/lib/table';
-import deleteUser from 'api/user/deleteUser';
-import editUserApi from 'api/user/editUser';
-import listUsers from 'api/user/listUsers';
-import updateRole from 'api/user/updateRole';
+import deleteUser from 'api/user/crud/delete';
+import listUsers from 'api/user/crud/getAll';
+import editUserApi from 'api/user/crud/update';
 import { ResizeTable } from 'components/ResizeTable';
 import { DATE_TIME_FORMATS } from 'constants/dateTimeFormats';
 import dayjs from 'dayjs';
@@ -124,9 +123,11 @@ function UserFunction({
 					userId: id,
 					name: updatedName,
 				}),
-				updateRole({
-					group_name: role,
+				editUserApi({
+					// TODO[vikrantgupta25]: fix this
+					// group_name: role,
 					userId: id,
+					name: role,
 				}),
 			]);
