+++ frontend/src/hooks/user/useGetUser.ts
@@ -1,14 +1,14 @@
-import getUser from 'api/user/getUser';
+import getUser from 'api/user/crud/get';
 import { useQuery, UseQueryResult } from 'react-query';
 import { SuccessResponseV2 } from 'types/api';
 import APIError from 'types/api/error';
 import { UserResponse } from 'types/api/user/getUser';
 
-const useGetUser = (userId: string, isLoggedIn: boolean): UseGetUser =>
+const useGetUser = (isLoggedIn: boolean): UseGetUser =>
 	useQuery({
-		queryFn: () => getUser({ userId }),
-		queryKey: [userId],
-		enabled: !!userId && !!isLoggedIn,
+		queryFn: () => getUser(),
+		queryKey: ['getMeUser'],
+		enabled: !!isLoggedIn,
 	});
 
 type UseGetUser = UseQueryResult<SuccessResponseV2<UserResponse>, APIError>;
